Improvements #7
