+++ MVVM/Model/Category.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TODO.MVVM.Model
+{
+    /// <summary>
+    /// Represents the category of a to-do item.
+    /// </summary>
+    public class Category
+    {
+        public string Name { get; set; }
+    }
+}
